+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - conversations summary extraction pipeline
+- docs and tests for init command
+- docs and tests for automated-import command
 
 ## [0.2.2] 2024-06-26